Added functionality for specifying taggers over types.
Created a class TypedToken that contains the Lemmatized<ChunkedToken> and the types that intersect with it.
Changed the RegexPattern wrapper in PatternTagger to handle TypedToken rather than Lemmatized<ChunkedToken>.
Implemented Logic for evaluating a regular expression with type values over some TypedTokens.
Added a method in the Tagger abstract class to allow PatternTagger to take as input the Types on the sentence accumulated from previous taggers.